ortalib
Enum Joker

pub enum Joker {

    Joker,
    JollyJoker,
    ZanyJoker,
    MadJoker,
    CrazyJoker,
    DrollJoker,
    SlyJoker,
    WilyJoker,
    CleverJoker,
    DeviousJoker,
    CraftyJoker,
    AbstractJoker,
    RaisedFist,
    Blackboard,
    Baron,
    GreedyJoker,
    LustyJoker,
    WrathfulJoker,
    GluttonousJoker,
    Fibonacci,
    ScaryFace,
    EvenSteven,
    OddTodd,
    Photograph,
    SmileyFace,
    FlowerPot,
    FourFingers,
    Shortcut,
    Mime,
    Pareidolia,
    Splash,
    SockAndBuskin,
    SmearedJoker,
    Blueprint,

}

The various Joker effects supported in Ortalab. Check the enum variant documentation for details on each of their individual effects.

Variants

Joker

+4 Mult.

Activates: Independent

JollyJoker

+8 Mult if played hand contains a Pair.

Activates: Independent

ZanyJoker

+12 Mult if played hand contains a Three of a Kind.

Activates: Independent

MadJoker

+10 Mult if played hand contains a Two Pair.

Activates: Independent

CrazyJoker

+12 Mult if played hand contains a Straight.

Activates: Independent

DrollJoker

+10 Mult if played hand contains a Flush.

Activates: Independent

SlyJoker

+50 Chips if played hand contains a Pair.

Activates: Independent

WilyJoker

+100 Chips if played hand contains a Three of a Kind.

Activates: Independent

CleverJoker

+80 Chips if played hand contains a Two Pair.

Activates: Independent

DeviousJoker

+100 Chips if played hand contains a Straight.

Activates: Independent

CraftyJoker

+80 Chips if played hand contains a Flush.

Activates: Independent

AbstractJoker

+3 Mult for each Joker card.

Activates: Independent

RaisedFist

Adds double the Rank::rank_value of lowest ranked card held in hand to Mult.

Activates: On Held

Blackboard

x3 Mult if all cards held in hand are ♠Spades or ♣Clubs.

Activates: Independent

Baron

Activates: On Held

GreedyJoker

Played cards with ♦Diamond suit give +3 Mult when scored.

Activates: On Scored

LustyJoker

Played cards with ♥Heart suit give +3 Mult when scored.

Activates: On Scored

WrathfulJoker

Played cards with ♠Spade suit give +3 Mult when scored.

Activates: On Scored

GluttonousJoker

Played cards with ♣Club suit give +3 Mult when scored.

Activates: On Scored

Fibonacci

Each played Ace, 2, 3, 5, or 8 gives +8 Mult when scored.

Activates: On Scored

ScaryFace

Played face cards give +30 Chips when scored.

Activates: On Scored

EvenSteven

Played cards with even rank give +4 Mult when scored.

Even ranks: (10, 8, 6, 4, 2)

Activates: On Scored

OddTodd

Played cards with odd rank give +31 Chips when scored

Odd ranks: (A, 9, 7, 5, 3)

Activates: On Scored

Photograph

First played face card gives x2 Mult when scored.

Activates: On Scored

SmileyFace

Played face cards give +5 Mult when scored.

Activates: On Scored

FlowerPot

x3 Mult if poker hand contains a ♦Diamond card, ♣Club card, ♥Heart card, and ♠Spade card.

Activates: Independent

FourFingers

All Flushes and Straights can be made with 4 cards.

Example 1: 4♦ 5♣ 6♦ 7♠ J♣ is a straight (J♣ not counted).

Example 2: 3♦ 7♦ 9♦ A♦ is a flush.

Example 3: 6♥ 7♥ 8♥ 9♥ is a straight flush.

Example 4: 3♥ 4♥ 5♣ 6♥ J♥ is a straight flush, since it contains both a straight and a flush.

Activates: N/A

Shortcut

Allows Straights to be made with gaps of 1 rank

(e.g. 10 8 6 5 3)

Activates: N/A

Mime

Retrigger all card held in hand abilities.

Activates: On Held

Pareidolia

All cards are considered face cards.

Activates: N/A

Splash

Every played card counts in scoring.

Activates: N/A

SockAndBuskin

Retrigger all played face cards.

Activates: On Scored

SmearedJoker

♥Hearts and ♦Diamonds count as the same suit, ♠Spades and ♣Clubs count as the same suit.

Activates: N/A

Blueprint

Copies the ability of Joker to the right.

Activates: Same as copied