Crate ortalib

Structs

Card

Represents one playing card.

JokerCard

Represents one Joker card. Simply includes the Joker effect, and a possible edition.

Round

Represents the total inputs of one scoring round in Ortalab.

Enums

Edition

Editions give Cards and Jokers a bonus when scoring them.

Enhancement

Editions give Cards a bonus when scoring them.

Joker

The various Joker effects supported in Ortalab. Check the enum variant documentation for details on each of their individual effects.

PokerHand

Poker Hands are sets of between one and five cards that can be played in Ortalab to obtain Chips and Mult for scoring.

Rank

Card Ranks are the thirteen values that can be found on a playing card in Ortalab. These values are Aces, the numerals 2 to 10, and the three face cards: Jacks, Queens and Kings. The value of each card can be accessed with Rank::rank_value as follows:

Suit

In Ortalab, there are 4 suits: ♥Hearts, ♣Clubs, ♦Diamonds, ♠Spades.

SuitColor

Represents the color of a Suit.

Type Aliases

Chips

Mult